cleric npc ai, fixes

+++ source/system/UnitData.cpp
@@ -201,7 +201,7 @@ bool TraderInfo::CanBuySell(const Item* item)
 		if(item->type == IT_CONSUMABLE)
 		{
 			const Consumable* c = static_cast<const Consumable*>(item);
-			if(IsSet(buy_consumable_flags, (1 << c->cons_type)))
+			if(IsSet(buy_consumable_flags, (1 << (int)c->cons_type)))
 				return true;
 		}
 		else
